make selected wallet the main view
