+++ packages/shared/lib/router.ts
@@ -1,15 +1,6 @@
 import { cleanupSignup, login, strongholdPassword, walletPin } from 'shared/lib/app'
 import { activeProfile, profiles } from 'shared/lib/profile'
-import {
-    AccountRoutes,
-    AppRoute,
-    LedgerRoutes,
-    SettingsRoutes,
-    SetupType,
-    Tabs,
-    WalletRoutes,
-} from 'shared/lib/typings/routes'
-import { selectedAccountId } from 'shared/lib/wallet'
+import { AccountRoutes, AppRoute, LedgerRoutes, SettingsRoutes, SetupType, Tabs } from 'shared/lib/typings/routes'
 import { get, readable, writable } from 'svelte/store'
 import { deepLinkRequestActive } from './deepLinking/deepLinking'
 import { closePopup } from './popup'
@@ -79,11 +70,6 @@ export const ledgerRoute = writable<LedgerRoutes>(LedgerRoutes.LegacyIntro)
  * Ledger setup routing history
  */
 export const ledgerRouteHistory = writable<string[]>([])
-
-/**
- * Wallet view route
- */
-export const walletRoute = writable<WalletRoutes>(WalletRoutes.Init)
 
 /**
  * Account view route
@@ -293,7 +279,6 @@ export const resetRouter = (): void => {
         setRoute(AppRoute.Welcome)
     }
 
-    walletRoute.set(WalletRoutes.Init)
     accountRoute.set(AccountRoutes.Init)
     settingsRoute.set(SettingsRoutes.Init)
     dashboardRoute.set(Tabs.Wallet)
@@ -302,9 +287,7 @@ export const resetRouter = (): void => {
 
 export const resetWalletRoute = (): void => {
     dashboardRoute.set(Tabs.Wallet)
-    walletRoute.set(WalletRoutes.Init)
     accountRoute.set(AccountRoutes.Init)
-    selectedAccountId.set(null)
 }
 
 export const resetLedgerRoute = (): void => {
