+++ packages/shared/lib/typings/routes.ts
@@ -40,6 +40,7 @@ export enum LedgerRoutes {
     AccountIndex = 'accountIndex',
 }
 
+// TODO: remove this
 export enum WalletRoutes {
     Init = 'init',
     Account = 'account',
