+++ packages/shared/routes/dashboard/wallet/views/ManageAccount.svelte
@@ -1,8 +1,8 @@
 <script lang="typescript">
     import { Button, Input, Text } from 'shared/components'
     import { getTrimmedLength } from 'shared/lib/helpers'
-    import { accountRoute, walletRoute } from 'shared/lib/router'
-    import { AccountRoutes, WalletRoutes } from 'shared/lib/typings/routes'
+    import { accountRoute } from 'shared/lib/router'
+    import { AccountRoutes } from 'shared/lib/typings/routes'
     import { api, MAX_ACCOUNT_NAME_LENGTH, selectedAccountId, wallet } from 'shared/lib/wallet'
     import { Locale } from 'shared/lib/typings/i18n'
     import { WalletAccount } from 'shared/lib/typings/wallet'
@@ -23,8 +23,8 @@
     const handleSaveClick = () => {
         const trimmedAccountAlias = accountAlias.trim()
         if (trimmedAccountAlias === alias) {
-            selectedAccountId.set(null)
-            walletRoute.set(WalletRoutes.Init)
+            // TODO: double check if we do want this change
+            accountRoute.set(AccountRoutes.Init)
             return
         }
         if (trimmedAccountAlias) {
@@ -57,8 +57,8 @@
                     }))
 
                     isBusy = false
-                    selectedAccountId.set(null)
-                    walletRoute.set(WalletRoutes.Init)
+                    // TODO: double check if we do want this change
+                    accountRoute.set(AccountRoutes.Init)
                 },
                 onError(err) {
                     isBusy = false
