+++ packages/shared/routes/dashboard/wallet/views/Receive.svelte
@@ -5,8 +5,8 @@
     import type { AccountIdentifier } from 'shared/lib/typings/account'
     import type { Locale } from 'shared/lib/typings/i18n'
     import { isLedgerProfile } from 'shared/lib/profile'
-    import { accountRoute, walletRoute } from 'shared/lib/router'
-    import { AccountRoutes, WalletRoutes } from 'shared/lib/typings/routes'
+    import { accountRoute } from 'shared/lib/router'
+    import { AccountRoutes } from 'shared/lib/typings/routes'
     import { setClipboard } from 'shared/lib/utils'
     import { hasGeneratedALedgerReceiveAddress, isSyncing } from 'shared/lib/wallet'
     import type { WalletAccount } from 'shared/lib/typings/wallet'
@@ -31,7 +31,6 @@
         onGenerateAddress(selectedAccount.id)
     }
     const handleCloseClick = () => {
-        walletRoute.set(WalletRoutes.Init)
         accountRoute.set(AccountRoutes.Init)
     }
 </script>
