+++ packages/shared/routes/dashboard/wallet/views/index.js
@@ -1,6 +1,5 @@
 export { default as Receive } from './Receive.svelte'
 export { default as Send } from './Send.svelte'
-export { default as Account } from './Account.svelte'
 export { default as CreateAccount } from './CreateAccount.svelte'
 export { default as ManageAccount } from './ManageAccount.svelte'
 export { default as WalletBalance } from './WalletBalance.svelte'
